voice-interview-agent/requirements.txt
langchain
langchain-core
langchain-openai  # Use for OpenAI API
langchain-community 
pydantic>=2.0.0
python-dotenv
tiktoken
speechrecognition
PyAudio
pyttsx3